feat: improve toggle button behavior
made toggle button optional
improved toggle button position
refactor files according to responsibility
