+++ .devcontainer/devcontainer.json
@@ -1,10 +1,10 @@
 // See https://aka.ms/vscode-remote/devcontainer.json for format details.
 {
   "name": "Boilerplate Card Development",
-  "image": "ludeeus/container:monster",
+  "image": "ghcr.io/ludeeus/devcontainer/generic:latest",
   "context": "..",
   "appPort": ["5000:5000", "9123:8123"],
-  "postCreateCommand": "npm install",
+  "postCreateCommand": "pip install tzdata && apk add build-base && npm install",
   "runArgs": [
     "-v",
     "${env:HOME}${env:USERPROFILE}/.ssh:/tmp/.ssh" // This is added so you can push from inside the container
