+++ .github/workflows/build.yml
@@ -16,7 +16,7 @@ jobs:
       - uses: actions/checkout@v1
       - name: Build
         run: |
-          npm install
+          npm ci
           npm run build
       - name: HACS Action
         uses: "hacs/action@main"
