+++ .gitignore
@@ -2,6 +2,5 @@
 node_modules/
 node_modules
 /.rpt2_cache/
-package-lock.json
 /dist
 .DS_Store
